Create functions to cache the inverse of a matrix
Added functions makeCacheMatrix() and cacheSolve() that create a matrix
object that can cache its inverse and retrieve the cached inverse (if
available) or compute the inverse and set the cached inverse (if not
available).

A sample test script is included below the main body of the functions
(in comments) to describe how the function is used and intended
operation.